Struct rusoto_cognito_identity::CreateIdentityPoolInput

pub struct CreateIdentityPoolInput {
    pub allow_unauthenticated_identities: bool,
    pub cognito_identity_providers: Option<Vec<CognitoIdentityProvider>>,
    pub developer_provider_name: Option<String>,
    pub identity_pool_name: String,
    pub open_id_connect_provider_ar_ns: Option<Vec<String>>,
    pub saml_provider_ar_ns: Option<Vec<String>>,
    pub supported_login_providers: Option<HashMap<String, String>>,
}

Input to the CreateIdentityPool action.

Fields
allow_unauthenticated_identities: bool

TRUE if the identity pool supports unauthenticated logins.

cognito_identity_providers: Option<Vec<CognitoIdentityProvider>>

An array of Amazon Cognito Identity user pools and their client IDs.

developer_provider_name: Option<String>

The "domain" by which Cognito will refer to your users. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

Once you have set a developer provider name, you cannot change it. Please take care in setting this parameter.

identity_pool_name: String

A string that you provide.

open_id_connect_provider_ar_ns: Option<Vec<String>>

A list of OpendID Connect provider ARNs.

saml_provider_ar_ns: Option<Vec<String>>

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

supported_login_providers: Option<HashMap<String, String>>

Optional key:value pairs mapping provider names to provider app IDs.