Struct rusoto_cognito_identity::RoleMapping

pub struct RoleMapping {
    pub ambiguous_role_resolution: Option<String>,
    pub rules_configuration: Option<RulesConfigurationType>,
    pub type_: String,
}

A role mapping.

Fields
ambiguous_role_resolution: Option<String>

If you specify Token or Rules as the Type, AmbiguousRoleResolution is required.

Specifies the action to be taken if either no rules match the claim value for the Rules type, or there is no cognito:preferred_role claim and there are multiple cognito:roles matches for the Token type.

rules_configuration: Option<RulesConfigurationType>

The rules to be used for mapping users to roles.

If you specify Rules as the role mapping type, RulesConfiguration is required.

type_: String

The role mapping type. Token will use cognito:roles and cognito:preferred_role claims from the Cognito identity provider token to map groups to roles. Rules will attempt to match claims from the token to map to a role.