Trait rusoto_cognito_identity::CognitoIdentity

pub trait CognitoIdentity {
    fn create_identity_pool(
        &self, 
        input: CreateIdentityPoolInput
    ) -> RusotoFuture<IdentityPool, CreateIdentityPoolError>;

    fn delete_identities(
        &self, 
        input: DeleteIdentitiesInput
    ) -> RusotoFuture<DeleteIdentitiesResponse, DeleteIdentitiesError>;

    fn delete_identity_pool(
        &self, 
        input: DeleteIdentityPoolInput
    ) -> RusotoFuture<(), DeleteIdentityPoolError>;

    fn describe_identity(
        &self, 
        input: DescribeIdentityInput
    ) -> RusotoFuture<IdentityDescription, DescribeIdentityError>;

    fn describe_identity_pool(
        &self, 
        input: DescribeIdentityPoolInput
    ) -> RusotoFuture<IdentityPool, DescribeIdentityPoolError>;

    fn get_credentials_for_identity(
        &self, 
        input: GetCredentialsForIdentityInput
    ) -> RusotoFuture<GetCredentialsForIdentityResponse, GetCredentialsForIdentityError>;

    fn get_id(
        &self, 
        input: GetIdInput
    ) -> RusotoFuture<GetIdResponse, GetIdError>;

    fn get_identity_pool_roles(
        &self, 
        input: GetIdentityPoolRolesInput
    ) -> RusotoFuture<GetIdentityPoolRolesResponse, GetIdentityPoolRolesError>;

    fn get_open_id_token(
        &self, 
        input: GetOpenIdTokenInput
    ) -> RusotoFuture<GetOpenIdTokenResponse, GetOpenIdTokenError>;

    fn get_open_id_token_for_developer_identity(
        &self, 
        input: GetOpenIdTokenForDeveloperIdentityInput
    ) -> RusotoFuture<GetOpenIdTokenForDeveloperIdentityResponse, GetOpenIdTokenForDeveloperIdentityError>;

    fn list_identities(
        &self, 
        input: ListIdentitiesInput
    ) -> RusotoFuture<ListIdentitiesResponse, ListIdentitiesError>;

    fn list_identity_pools(
        &self, 
        input: ListIdentityPoolsInput
    ) -> RusotoFuture<ListIdentityPoolsResponse, ListIdentityPoolsError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceInput
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn lookup_developer_identity(
        &self, 
        input: LookupDeveloperIdentityInput
    ) -> RusotoFuture<LookupDeveloperIdentityResponse, LookupDeveloperIdentityError>;

    fn merge_developer_identities(
        &self, 
        input: MergeDeveloperIdentitiesInput
    ) -> RusotoFuture<MergeDeveloperIdentitiesResponse, MergeDeveloperIdentitiesError>;

    fn set_identity_pool_roles(
        &self, 
        input: SetIdentityPoolRolesInput
    ) -> RusotoFuture<(), SetIdentityPoolRolesError>;

    fn tag_resource(
        &self, 
        input: TagResourceInput
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn unlink_developer_identity(
        &self, 
        input: UnlinkDeveloperIdentityInput
    ) -> RusotoFuture<(), UnlinkDeveloperIdentityError>;

    fn unlink_identity(
        &self, 
        input: UnlinkIdentityInput
    ) -> RusotoFuture<(), UnlinkIdentityError>;

    fn untag_resource(
        &self, 
        input: UntagResourceInput
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_identity_pool(
        &self, 
        input: IdentityPool
    ) -> RusotoFuture<IdentityPool, UpdateIdentityPoolError>;
}

Trait representing the capabilities of the Amazon Cognito Identity API. Amazon Cognito Identity clients implement this trait.

Required methods

fn create_identity_pool(
&self,
input: CreateIdentityPoolInput
) -> RusotoFuture<IdentityPool, CreateIdentityPoolError>

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. The limit on identity pools is 60 per account. The keys for SupportedLoginProviders are as follows:

Facebook: graph.facebook.com

Google: accounts.google.com

Amazon: www.amazon.com

Twitter: api.twitter.com

Digits: www.digits.com

You must use AWS Developer credentials to call this API.

fn delete_identities(
&self,
input: DeleteIdentitiesInput
) -> RusotoFuture<DeleteIdentitiesResponse, DeleteIdentitiesError>

Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

You must use AWS Developer credentials to call this API.

fn delete_identity_pool(
&self,
input: DeleteIdentityPoolInput
) -> RusotoFuture<(), DeleteIdentityPoolError>

Deletes an identity pool. Once a pool is deleted, users will not be able to authenticate with the pool.

You must use AWS Developer credentials to call this API.

fn describe_identity(
&self,
input: DescribeIdentityInput
) -> RusotoFuture<IdentityDescription, DescribeIdentityError>

Returns metadata related to the given identity, including when the identity was created and any associated linked logins.

You must use AWS Developer credentials to call this API.

fn describe_identity_pool(
&self,
input: DescribeIdentityPoolInput
) -> RusotoFuture<IdentityPool, DescribeIdentityPoolError>

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.

You must use AWS Developer credentials to call this API.

fn get_credentials_for_identity(
&self,
input: GetCredentialsForIdentityInput
) -> RusotoFuture<GetCredentialsForIdentityResponse, GetCredentialsForIdentityError>

Returns credentials for the provided identity ID. Any provided logins will be validated against supported login providers. If the token is for cognito-identity.amazonaws.com, it will be passed through to AWS Security Token Service with the appropriate role for the token.

This is a public API. You do not need any credentials to call this API.

fn get_id(&self, input: GetIdInput) -> RusotoFuture<GetIdResponse, GetIdError>

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.

This is a public API. You do not need any credentials to call this API.

fn get_identity_pool_roles(
&self,
input: GetIdentityPoolRolesInput
) -> RusotoFuture<GetIdentityPoolRolesResponse, GetIdentityPoolRolesError>

Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.

fn get_open_id_token(
&self,
input: GetOpenIdTokenInput
) -> RusotoFuture<GetOpenIdTokenResponse, GetOpenIdTokenError>

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenId token is valid for 10 minutes.

This is a public API. You do not need any credentials to call this API.

fn get_open_id_token_for_developer_identity(
&self,
input: GetOpenIdTokenForDeveloperIdentityInput
) -> RusotoFuture<GetOpenIdTokenForDeveloperIdentityResponse, GetOpenIdTokenForDeveloperIdentityError>

Registers (or retrieves) a Cognito IdentityId and an OpenID Connect token for a user authenticated by your backend authentication process. Supplying multiple logins will create an implicit linked account. You can only specify one developer provider as part of the Logins map, which is linked to the identity pool. The developer provider is the "domain" by which Cognito will refer to your users.

You can use GetOpenIdTokenForDeveloperIdentity to create a new identity and to link new logins (that is, user credentials issued by a public provider or developer provider) to an existing identity. When you want to create a new identity, the IdentityId should be null. When you want to associate a new login with an existing authenticated/unauthenticated identity, you can do so by providing the existing IdentityId. This API will create the identity in the specified IdentityPoolId.

You must use AWS Developer credentials to call this API.

fn list_identities(
&self,
input: ListIdentitiesInput
) -> RusotoFuture<ListIdentitiesResponse, ListIdentitiesError>

Lists the identities in an identity pool.

You must use AWS Developer credentials to call this API.

fn list_identity_pools(
&self,
input: ListIdentityPoolsInput
) -> RusotoFuture<ListIdentityPoolsResponse, ListIdentityPoolsError>

Lists all of the Cognito identity pools registered for your account.

You must use AWS Developer credentials to call this API.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceInput
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists the tags that are assigned to an Amazon Cognito identity pool.

A tag is a label that you can apply to identity pools to categorize and manage them in different ways, such as by purpose, owner, environment, or other criteria.

You can use this action up to 10 times per second, per account.

fn lookup_developer_identity(
&self,
input: LookupDeveloperIdentityInput
) -> RusotoFuture<LookupDeveloperIdentityResponse, LookupDeveloperIdentityError>

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifier values associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID. If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

LookupDeveloperIdentity is intended for low-throughput control plane operations: for example, to enable customer service to locate an identity ID by username. If you are using it for higher-volume operations such as user authentication, your requests are likely to be throttled. GetOpenIdTokenForDeveloperIdentity is a better option for higher-volume operations for user authentication.

You must use AWS Developer credentials to call this API.

fn merge_developer_identities(
&self,
input: MergeDeveloperIdentitiesInput
) -> RusotoFuture<MergeDeveloperIdentitiesResponse, MergeDeveloperIdentitiesError>

Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider. You can use this action to request that discrete users be merged and identified as a single user in the Cognito environment. Cognito associates the given source user (SourceUserIdentifier) with the IdentityId of the DestinationUserIdentifier. Only developer-authenticated users can be merged. If the users to be merged are associated with the same public provider, but as two different users, an exception will be thrown.

The number of linked logins is limited to 20. So, the number of linked logins for the source user, SourceUserIdentifier, and the destination user, DestinationUserIdentifier, together should not be larger than 20. Otherwise, an exception will be thrown.

You must use AWS Developer credentials to call this API.

fn set_identity_pool_roles(
&self,
input: SetIdentityPoolRolesInput
) -> RusotoFuture<(), SetIdentityPoolRolesError>

Sets the roles for an identity pool. These roles are used when making calls to GetCredentialsForIdentity action.

You must use AWS Developer credentials to call this API.

fn tag_resource(
&self,
input: TagResourceInput
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Assigns a set of tags to an Amazon Cognito identity pool. A tag is a label that you can use to categorize and manage identity pools in different ways, such as by purpose, owner, environment, or other criteria.

Each tag consists of a key and value, both of which you define. A key is a general category for more specific values. For example, if you have two versions of an identity pool, one for testing and another for production, you might assign an Environment tag key to both identity pools. The value of this key might be Test for one identity pool and Production for the other.

Tags are useful for cost tracking and access control. You can activate your tags so that they appear on the Billing and Cost Management console, where you can track the costs associated with your identity pools. In an IAM policy, you can constrain permissions for identity pools based on specific tags or tag values.

You can use this action up to 5 times per second, per account. An identity pool can have as many as 50 tags.

fn unlink_developer_identity(
&self,
input: UnlinkDeveloperIdentityInput
) -> RusotoFuture<(), UnlinkDeveloperIdentityError>

Unlinks a DeveloperUserIdentifier from an existing identity. Unlinked developer users will be considered new identities next time they are seen. If, for a given Cognito identity, you remove all federated identities as well as the developer user identifier, the Cognito identity becomes inaccessible.

You must use AWS Developer credentials to call this API.

fn unlink_identity(
&self,
input: UnlinkIdentityInput
) -> RusotoFuture<(), UnlinkIdentityError>

Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.

This is a public API. You do not need any credentials to call this API.

fn untag_resource(
&self,
input: UntagResourceInput
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Removes the specified tags from an Amazon Cognito identity pool. You can use this action up to 5 times per second, per account

fn update_identity_pool(
&self,
input: IdentityPool
) -> RusotoFuture<IdentityPool, UpdateIdentityPoolError>

Updates an identity pool.

You must use AWS Developer credentials to call this API.

Loading content...