Struct rusoto_cognito_identity::CognitoIdentityClient

pub struct CognitoIdentityClient { /* fields omitted */ }

A client for the Amazon Cognito Identity API.

Methods

impl CognitoIdentityClient

pub fn new(region: Region) -> CognitoIdentityClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CognitoIdentityClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,